requirements.txt
python-telegram-bot==21.7
aiohttp==3.10.11
tenacity==9.0.0